+++ application/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@tailwindcss/vite": "^4.1.13",
+    "@tanstack/react-query": "^5.90.2",
     "@types/express": "^5.0.3",
     "@types/node": "^24.5.2",
     "express": "^5.1.0",
